cts: Merge CtsOptions state into TritonCTS
Removes CtsOptions global state anti-pattern from TritonCTS.
Migrates all CtsOptions fields and getters/setters directly into the
TritonCTS class. This aligns the architecture with idiomatic OpenROAD
patterns, avoiding disjoint parameter objects where unnecessary.
